# Remove global variable Main.target 