--- a/src/algorithms/precondition.jl
+++ b/src/algorithms/precondition.jl
@@ -8,8 +8,6 @@ using CUDA
 using CUDA.CUSPARSE
 using TimerOutputs
 
-using ..ExaPF: Kernels
-
 cuzeros = CUDA.zeros
 
 abstract type AbstractPreconditioner end
@@ -77,24 +75,23 @@ mutable struct Preconditioner <: AbstractPreconditioner
             end
         end
         P = sparse(row, col, nzval)
-        # TODO: clean global variables
         if Main.target == "cuda"
-            global cupartitions = Vector{CuVector{Int64}}(undef, npart)
+            cupartitions = Vector{CuVector{Int64}}(undef, npart)
             for i in 1:npart
                 cupartitions[i] = CuVector{Int64}(partitions[i])
             end
-            global cuJs = Vector{CuMatrix{Float64}}(undef, length(partitions))
+            cuJs = Vector{CuMatrix{Float64}}(undef, length(partitions))
             for i in 1:length(partitions)
                 cuJs[i] = CuMatrix{Float64}(I, nJs, nJs)
             end
-            global cumap = cu(map)
-            global cupart = cu(part)
-            global P = CuSparseMatrixCSR(P)
+            cumap = cu(map)
+            cupart = cu(part)
+            P = CuSparseMatrixCSR(P)
         else
-            global cuJs = nothing
-            global cupartitions = nothing
-            global cumap = nothing
-            global cupart = nothing
+            cuJs = nothing
+            cupartitions = nothing
+            cumap = nothing
+            cupart = nothing
         end
         return new(npart, nJs, partitions, cupartitions, Js, cuJs, map, cumap, part, cupart, P)
     end
@@ -123,9 +120,8 @@ function build_adjmatrix(A)
 end
 
 function fillblock_gpu!(cuJs, partition, map, rowPtr, colVal, nzVal, part, b)
-
-    Kernels.@getstrideindex()
-
+    index = (blockIdx().x - 1) * blockDim().x + threadIdx().x
+    stride = blockDim().x * gridDim().x
     for i in index:stride:length(partition)
         for j in rowPtr[partition[i]]:rowPtr[partition[i]+1]-1
             if b == part[colVal[j]]
@@ -137,9 +133,8 @@ function fillblock_gpu!(cuJs, partition, map, rowPtr, colVal, nzVal, part, b)
 end
 
 function fillP_gpu!(cuJs, partition, map, rowPtr, colVal, nzVal, part, b)
-
-    Kernels.@getstrideindex()
-
+    index = (blockIdx().x - 1) * blockDim().x + threadIdx().x
+    stride = blockDim().x * gridDim().x
     for i in index:stride:length(partition)
         for j in rowPtr[partition[i]]:rowPtr[partition[i]+1]-1
             if b == part[colVal[j]]
@@ -150,55 +145,57 @@ function fillP_gpu!(cuJs, partition, map, rowPtr, colVal, nzVal, part, b)
     return nothing
 end
 
-function update(J, p, to)
+function update(J::CuSparseMatrixCSR, p, to)
     m = size(J, 1)
     n = size(J, 2)
     nblocks = length(p.partitions)
-    if J isa CuSparseMatrixCSR
-        @timeit to "Fill Block Jacobi" begin
-            Kernels.@sync begin
-                for b in 1:nblocks
-                    Kernels.@dispatch threads=16 blocks=16 fillblock_gpu!(p.cuJs[b], p.cupartitions[b], p.cumap, J.rowPtr, J.colVal, J.nzVal, p.cupart, b)
-                end
-            end
-        end
-        @timeit to "Invert blocks" begin
-            CUDA.@sync pivot, info = CUDA.CUBLAS.getrf_batched!(p.cuJs, true)
-            CUDA.@sync pivot, info, p.cuJs = CUDA.CUBLAS.getri_batched(p.cuJs, pivot)
-        end
-        p.P.nzVal .= 0.0
-        @timeit to "Move blocks to P" begin
-            Kernels.@sync begin
-                for b in 1:nblocks
-                    Kernels.@dispatch threads=16 blocks=16 fillP_gpu!(p.cuJs[b], p.cupartitions[b], p.cumap, p.P.rowPtr, p.P.colVal, p.P.nzVal, p.cupart, b)
-                end
-            end
-        end
-    else
-        @timeit to "Fill Block Jacobi" begin
+    @timeit to "Fill Block Jacobi" begin
+        CUDA.@sync begin
             for b in 1:nblocks
-                for i in p.partitions[b]
-                    for j in J.colptr[i]:J.colptr[i+1]-1
-                        if b == p.part[J.rowval[j]]
-                            p.Js[b][p.map[J.rowval[j]], p.map[i]] = J.nzval[j]
-                        end
+                @cuda threads=16 blocks=16 fillblock_gpu!(p.cuJs[b], p.cupartitions[b], p.cumap, J.rowPtr, J.colVal, J.nzVal, p.cupart, b)
+            end
+        end
+    end
+    @timeit to "Invert blocks" begin
+        CUDA.@sync pivot, info = CUDA.CUBLAS.getrf_batched!(p.cuJs, true)
+        CUDA.@sync pivot, info, p.cuJs = CUDA.CUBLAS.getri_batched(p.cuJs, pivot)
+    end
+    p.P.nzVal .= 0.0
+    @timeit to "Move blocks to P" begin
+        CUDA.@sync begin
+            for b in 1:nblocks
+                @cuda threads=16 blocks=16 fillP_gpu!(p.cuJs[b], p.cupartitions[b], p.cumap, p.P.rowPtr, p.P.colVal, p.P.nzVal, p.cupart, b)
+            end
+        end
+    end
+    return p.P
+end
+
+function update(J::SparseMatrixCSC, p, to)
+    nblocks = length(p.partitions)
+    @timeit to "Fill Block Jacobi" begin
+        for b in 1:nblocks
+            for i in p.partitions[b]
+                for j in J.colptr[i]:J.colptr[i+1]-1
+                    if b == p.part[J.rowval[j]]
+                        p.Js[b][p.map[J.rowval[j]], p.map[i]] = J.nzval[j]
                     end
                 end
             end
         end
-        @timeit to "Invert blocks" begin
-            for b in 1:nblocks
-                p.Js[b] = inv(p.Js[b])
-            end
+    end
+    @timeit to "Invert blocks" begin
+        for b in 1:nblocks
+            p.Js[b] = inv(p.Js[b])
         end
-        p.P.nzval .= 0.0
-        @timeit to "Move blocks to P" begin
-            for b in 1:nblocks
-                for i in p.partitions[b]
-                    for j in p.P.colptr[i]:p.P.colptr[i+1]-1
-                        if b == p.part[p.P.rowval[j]]
-                            p.P.nzval[j] += p.Js[b][p.map[p.P.rowval[j]], p.map[i]]
-                        end
+    end
+    p.P.nzval .= 0.0
+    @timeit to "Move blocks to P" begin
+        for b in 1:nblocks
+            for i in p.partitions[b]
+                for j in p.P.colptr[i]:p.P.colptr[i+1]-1
+                    if b == p.part[p.P.rowval[j]]
+                        p.P.nzval[j] += p.Js[b][p.map[p.P.rowval[j]], p.map[i]]
                     end
                 end
             end