--- a/src/iterative.jl
+++ b/src/iterative.jl
@@ -17,11 +17,12 @@ cuzeros = CUDA.zeros
 bicgstab according to
 
 Van der Vorst, Henk A.
-"Bi-CGSTAB: A fast and smoothly converging variant of Bi-CG for the solution of nonsymmetric linear systems."
+"Bi-CGSTAB: A fast and smoothly converging variant of Bi-CG for the solution
+of nonsymmetric linear systems."
 SIAM Journal on scientific and Statistical Computing 13, no. 2 (1992): 631-644.
 """
-function bicgstab(A, b, P, xi, to::TimerOutput; tol = 1e-6, maxiter = size(A,1),
-                  verbose=false)
+function bicgstab(A, b, P, xi, to::TimerOutput;
+                  tol=1e-6, maxiter=size(A, 1), verbose=false)
     # parameters
     n    = size(b, 1)
     x0   = similar(b)
@@ -32,9 +33,11 @@ function bicgstab(A, b, P, xi, to::TimerOutput; tol = 1e-6, maxiter = size(A,1),
     alpha = 1.0
     omega0 = 1.0
     if A isa SparseArrays.SparseMatrixCSC
-        v0   = p0 = zeros(Float64, n)
+        v0 = zeros(Float64, n)
+        p0 = zeros(Float64, n)
     else
-        v0   = p0 = cuzeros(Float64, n)
+        v0 = cuzeros(Float64, n)
+        p0 = cuzeros(Float64, n)
     end
 
     ri     = copy(r0)