This PR migrates ExaPF to KernelAbstractions.jl, using the `@kernel` macro
- `residualFunction_polar!` is rewritten with minor optimization 
- in `ad.jl`, `myseed` and `getpartials` are also passed to `@kernel`, and magically ... it works perfectly fine with ForwardDiff's dual numbers
- `precondition.jl` is also updated to remove i) the global variables ii) remove the previous `Kernels`'s  macro

This PR adds also a small change to the interface. To solve the PF on the GPU, one has to write from now on:
```julia
solve(pf, nblocks, precond, device=CUDADevice())
```
By default, `solve` runs on the CPU (`CPU()` device)

It still remains to do some benchmarks with the previous version. I am planning to use the script `test/perf.jl` for that, using the 70K example.