@frapac 2020-07-30
Here is a benchmark comparing `master` with this PR (using the script `test/perf.jl`).

On `case14.raw`, we get:
```
[new version] KernelAbstractions
[CPU] Benchmark function residualFunction_polar!
  2.872 μs (18 allocations: 1.05 KiB)
[CPU] Benchmark residualJacobianAD!
  41.965 μs (410 allocations: 14.63 KiB)
[GPU] Benchmark residualFunction_polar!
  48.430 μs (217 allocations: 8.47 KiB)
[GPU] Benchmark residualJacobianAD!
  296.646 μs (853 allocations: 29.53 KiB)

[old version] ExaPF.Kernels
[CPU] Benchmark function residualFunction_polar!
  1.208 μs (0 allocations: 0 bytes)
[CPU] Benchmark residualJacobianAD!
  31.812 μs (352 allocations: 11.48 KiB)
[GPU] Benchmark residualFunction_polar!
  6.573 ms (2790 allocations: 115.61 KiB)
[GPU] Benchmark residualJacobianAD!
  12.373 ms (6393 allocations: 306.06 KiB)
```

And with `ACTIVSg70K`, we get: 
```
[new version] KernelAbstractions
[CPU] Benchmark function residualFunction_polar!
  3.758 ms (21 allocations: 1.09 KiB)
[CPU] Benchmark residualJacobianAD!
  287.889 ms (5570899 allocations: 141.29 MiB)
[GPU] Benchmark residualFunction_polar!
  130.356 μs (376 allocations: 10.95 KiB)
[GPU] Benchmark residualJacobianAD!
  6.162 ms (4505 allocations: 86.59 KiB)

[old version] ExaPF.Kernels
[CPU] Benchmark function residualFunction_polar!
  7.956 ms (1 allocation: 16 bytes)
[CPU] Benchmark residualJacobianAD!
  296.346 ms (5570834 allocations: 141.29 MiB)
[GPU] Benchmark residualFunction_polar!
  36.433 s (14840943 allocations: 603.88 MiB)
[GPU] Benchmark residualJacobianAD!
  72.70 s (2.76 MiB)
```